requirements.txt
aiohttp
requests
discord.py[voice]~=1.7.3
markdownify
PyYAML
PyNaCl
pytube
validators
shortuuid